streamlit_ornek/requirements.txt
openai==1.3.7
langchain==0.0.345
tiktoken==0.5.2
cohere==4.37
python-dotenv==1.0.0
streamlit==1.29.0 #pip install streamlit